+++ pystac/__init__.py
@@ -90,6 +90,7 @@ import pystac.extensions.file
 import pystac.extensions.grid
 import pystac.extensions.item_assets
 import pystac.extensions.label
+import pystac.extensions.mgrs
 import pystac.extensions.pointcloud
 import pystac.extensions.projection
 import pystac.extensions.raster
@@ -110,6 +111,7 @@ EXTENSION_HOOKS = pystac.extensions.hooks.RegisteredExtensionHooks(
         pystac.extensions.grid.GRID_EXTENSION_HOOKS,
         pystac.extensions.item_assets.ITEM_ASSETS_EXTENSION_HOOKS,
         pystac.extensions.label.LABEL_EXTENSION_HOOKS,
+        pystac.extensions.mgrs.MGRS_EXTENSION_HOOKS,
         pystac.extensions.pointcloud.POINTCLOUD_EXTENSION_HOOKS,
         pystac.extensions.projection.PROJECTION_EXTENSION_HOOKS,
         pystac.extensions.raster.RASTER_EXTENSION_HOOKS,