+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v1.7.3]
+
 ### Fixed
 
 - Duplicate `self` links in Items ([#1103](https://github.com/stac-utils/pystac/pull/1103))
@@ -680,7 +682,8 @@ use `Band.create`
 
 Initial release.
 
-[Unreleased]: <https://github.com/stac-utils/pystac/compare/v1.7.2..main>
+[Unreleased]: <https://github.com/stac-utils/pystac/compare/v1.7.3..main>
+[v1.7.3]: <https://github.com/stac-utils/pystac/compare/v1.7.2..v1.7.3>
 [v1.7.2]: <https://github.com/stac-utils/pystac/compare/v1.7.1..v1.7.2>
 [v1.7.1]: <https://github.com/stac-utils/pystac/compare/v1.7.0..v1.7.1>
 [v1.7.0]: <https://github.com/stac-utils/pystac/compare/v1.6.1..v1.7.0>